CommandsProcessor - intermediate

+++ DashCamGPSView/ZipOutput/Program.cs
@@ -3,7 +3,9 @@ using System.Collections.Generic;
 using System.IO;
 using System.IO.Compression;
 using System.Linq;
+using System.Reflection;
 using System.Text;
+using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
 namespace ZipOutput
@@ -20,27 +22,42 @@ namespace ZipOutput
                     return;
                 }
 
-                string DirectoryName = Path.GetFullPath(args[0]);
-                string ZipFileName = Path.GetFullPath(args[1]);
-
-                Console.WriteLine("Compressing: {0} into: {1}", DirectoryName, ZipFileName);
-
-                if(!Directory.Exists(DirectoryName))
+                if (args[0] == "zip")
                 {
-                    Console.WriteLine("Error: Directry does not exists: " + DirectoryName);
-                    return;
+                    string DirectoryName = Path.GetFullPath(args[1]);
+                    string ZipFileName = Path.GetFullPath(args[2]);
+
+                    Console.WriteLine("Compressing: {0} into: {1}", DirectoryName, ZipFileName);
+
+                    if (!Directory.Exists(DirectoryName))
+                    {
+                        Console.WriteLine("Error: Directry does not exists: " + DirectoryName);
+                        return;
+                    }
+
+                    if (File.Exists(ZipFileName))
+                        File.Delete(ZipFileName);
+
+                    string[] pdbFiles = Directory.GetFiles(DirectoryName, "*.pdb");
+                    foreach (string file in pdbFiles)
+                    {
+                        File.Delete(file);
+                    }
+
+                    ZipFile.CreateFromDirectory(DirectoryName, ZipFileName);
                 }
-
-                if (File.Exists(ZipFileName))
-                    File.Delete(ZipFileName);
-
-                string[] pdbFiles = Directory.GetFiles(DirectoryName, "*.pdb");
-                foreach (string file in pdbFiles)
+                else if (args[0] == "ver")
                 {
-                    File.Delete(file);
-                }
+                    string assemblyInfoFileName = args[1];
+                    if (!File.Exists(assemblyInfoFileName))
+                    {
+                        Console.WriteLine("Error: assemblyInfoFileName does not exists: " + assemblyInfoFileName);
+                        return;
+                    }
 
-                ZipFile.CreateFromDirectory(DirectoryName, ZipFileName);
+                    string assemblyInfoFileText = File.ReadAllText(assemblyInfoFileName);
+                    Match m = Regex.Match(assemblyInfoFileText, @"(?<major>\d{1,3})\.(?<minor>\d{1,3})\.(?<build>\d{1,3})\.(?<revision>\d{1,3})");
+                }
             }
             catch (Exception err)
             {
